requirements.txt
Wikipedia-API
beautifulsoup4
lxml
psycopg2-binary
future
yandexfreetranslate
PySocks
emoji
requests
sqlalchemy
python-telegram-bot==11.1.0
feedparser
